terraform/requirements.txt
functions-framework==3.*
requests
pandas
numpy
google-cloud-storage
